Convert TypeScript to Families
