+++ typescript/package.json
@@ -33,5 +33,27 @@
 	],
 	"engines": {
 		"node": ">=18.0.0"
+	},
+	"exports": {
+		".": {
+			"import": "./dist/esm/index.js",
+			"require": "./dist/cjs/index.js",
+			"types": "./dist/esm/index.d.ts"
+		},
+		"./terminology": {
+			"import": "./dist/esm/terminology.js",
+			"require": "./dist/cjs/terminology.js",
+			"types": "./dist/esm/terminology.d.ts"
+		},
+		"./convert": {
+			"import": "./dist/esm/convert.js",
+			"require": "./dist/cjs/convert.js",
+			"types": "./dist/esm/convert.d.ts"
+		},
+		"./insight": {
+			"import": "./dist/esm/insight.js",
+			"require": "./dist/cjs/insight.js",
+			"types": "./dist/esm/insight.d.ts"
+		}
 	}
 }
