+++ typescript/src/insight.ts
@@ -1,4 +1,5 @@
 import { Bundle, MeasureReport, Measure, Patient, RiskAssessment, OperationOutcome } from "fhir/r4";
+import { IHttpHandler } from "./httpHandler";
 
 const raSegments = [
   "community nondual aged",
@@ -24,3 +25,34 @@ export type InsightRiskProfileRequest = {
 export type InsightRiskProfileResource = Patient | MeasureReport | Measure | RiskAssessment | OperationOutcome;
 
 export type InsightRiskProfileResponse = Bundle<InsightRiskProfileResource>;
+
+export class InsightApi {
+  private httpHandler: IHttpHandler;
+
+  constructor(httpHandler: IHttpHandler) {
+    this.httpHandler = httpHandler;
+  }
+  /**
+   * Computes an HCC Risk Adjustment Profile for the provided patient
+   * @param request A FHIR R4 Bundle containing data for a single patient
+   * @returns A new FHIR R4 Bundle containing measure and assessment resources
+   * @link https://rosetta-api.docs.careevolution.com/insight/risk_profile.html
+   */
+  riskProfile(request: InsightRiskProfileRequest): Promise<InsightRiskProfileResponse> {
+    const parameters = new URLSearchParams();
+    if (request.hccVersion) {
+      parameters.append("hcc_version", request.hccVersion);
+    }
+    if (request.periodEndDate) {
+      parameters.append("period_end_date", request.periodEndDate);
+    }
+    if (request.raSegment) {
+      parameters.append("ra_segment", request.raSegment);
+    }
+    let route = "/insight/v1/riskprofile";
+    if (parameters.toString()) {
+      route += `?${parameters.toString()}`;
+    }
+    return this.httpHandler.post(route, request.content);
+  }
+}
